+++ ml/context/context.go
@@ -64,8 +64,8 @@ import (
 //		func ModelGraph(ctx *context.Context, inputs []*Node) (logits *Node) {
 //			...
 //			{
-//				ctx := ctx.In("output_layer)  // Enter "output_layer" scope in temporary new context (same data, different scope)
-//	        	ctx.SetParam("dropout_rate", 0.6  // Let's say we want the "output_layer" only to have dropout=0.6
+//				ctx := ctx.In("output_layer")  // Enter "output_layer" scope in temporary new context (same data, different scope)
+//	        	ctx.SetParam("dropout_rate", 0.6)  // Let's say we want the "output_layer" only to have dropout=0.6
 //	         	logits = Dense(ctx, logits, output_dim)
 //			}  // Exiting "output_later" scope, ctx is back to it's original scope.
 //		}
@@ -308,6 +308,21 @@ func (ctx *Context) GetParam(key string) (value any, found bool) {
 	return ctx.data.params.Get(ctx.scope, key)
 }
 
+// GetParamOr either returns the value for the given param key in the context `ctx`,
+// searching successively from the current scope back to the root scope ("/"), or if the
+// key is not found, returns the given default value.
+//
+// It casts the value to the given type, and it will panic is that fails.
+//
+// It's a convenience method around `ctx.GetParam`.
+func GetParamOr[T any](ctx *Context, key string, defaultValue T) T {
+	valueAny, found := ctx.GetParam(key)
+	if !found {
+		return defaultValue
+	}
+	return valueAny.(T)
+}
+
 // SetParam sets the given param in the current scope. It will be visible (by GetParam)
 // within this scope and descendant scopes (but not by other scopes).
 //
@@ -321,7 +336,7 @@ func (ctx *Context) EnumerateParams(fn func(scope, key string, value any)) {
 	ctx.data.params.Enumerate(fn)
 }
 
-// GetGraphParam returns, for the givne graph, the value for the given param key,
+// GetGraphParam returns the value for the given param key for the given graph,
 // searching successively from the current scope back to the root scope ("/"), in
 // case the key is not found.
 //
@@ -331,7 +346,7 @@ func (ctx *Context) EnumerateParams(fn func(scope, key string, value any)) {
 // This is very similar to GetParam, but used for parameters that are graph specific.
 // For example Context.IsTraining and Context.SetTraining uses a Graph parameter to
 // set this state, as the same Context is used for evaluation/inference graphs and
-// training graphs, and they should have different values.
+// training graphs, and they will have different values.
 func (ctx *Context) GetGraphParam(g *Graph, key string) (value any, found bool) {
 	var graphParams *ScopedParams
 	graphParams, found = ctx.data.graphParams[g.GraphId()]
@@ -341,9 +356,32 @@ func (ctx *Context) GetGraphParam(g *Graph, key string) (value any, found bool) 
 	return graphParams.Get(ctx.scope, key)
 }
 
+// GetGraphParamOr either returns the value for the given param key for the given graph,
+// searching successively from the current scope back to the root scope ("/"), or if the
+// key is not found, returns the given default value.
+//
+// It casts the value to the given type, and it will panic is that fails.
+//
+// It's a convenience method around `ctx.GetGraphParam`.
+//
+// This is very similar to GetParamOr, but used for parameters that are graph specific.
+// For example Context.IsTraining and Context.SetTraining uses a Graph parameter to
+// set this state, as the same Context is used for evaluation/inference graphs and
+// training graphs, and they will have different values.
+func GetGraphParamOr[T any](ctx *Context, g *Graph, key string, defaultValue T) T {
+	valueAny, found := ctx.GetParam(key)
+	if !found {
+		return defaultValue
+	}
+	return valueAny.(T)
+}
+
 // SetGraphParam sets the given Graph param in the current scope. It will be visible (by
 // GetGraphParam) for this Graph within this scope and descendant scopes (but not by other
 // scopes).
+//
+// Notice each time a new graph is created, the associated "graph parameters" in the context
+// will be empty.
 //
 // This is very similar to SetParam, but used for parameters that are graph specific.
 // For example Context.IsTraining and Context.SetTraining uses a Graph parameter to