+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Enable copy elision -- which makes `std::move` not necessary.
 * Temporarily copied `xla/mlir/utils` library to `third-party/xla_mlir`, since it is not available in all XLA distributions.
+* Added `context.GetParamOr`, using generics to cast to the desired type, and allowing a default value.
 
 ## v0.7.2 - 2023/10/27